requirements.txt
fastapi>=0.111.0,<0.112.0
starlette>=0.37.2,<0.38.0
uvicorn>=0.30.0,<0.31.0
sqlalchemy>=2.0.30,<2.1.0
psycopg2-binary>=2.9.9,<3.0.0
python-dotenv>=1.0.1,<1.1.0
pydantic>=1.10.14,<1.11.0
httpx>=0.27.0,<0.28.0
pytest>=8.2.2,<8.3.0